add webpack configuration
