+++ indecision-app/package.json
@@ -6,7 +6,8 @@
   "license": "MIT",
   "scripts": {
     "serve": "live-server public/",
-    "build": "babel src/app.js --out-file=public/scripts/app.js --presets=env,react --watch"
+    "build": "webpack --watch", 
+    "build-babel": "babel src/app.js --out-file=public/scripts/app.js --presets=env,react --watch"
   },
   "dependencies": {
     "babel-cli": "6.24.1",
