+++ indecision-app/public/index.html
@@ -6,8 +6,6 @@
     </head>
     <body>
         <div id="app"></div>
-        <script src="https://unpkg.com/react@16.0.0/umd/react.development.js"></script>
-        <script src="https://unpkg.com/react-dom@16.0.0/umd/react-dom.development.js"></script>   
-        <script src="./scripts/app.js"></script>  
+        <script src="./boundle.js"></script>  
     </body>
 </html>
